se reordeno las estructuras de las carpetas, se agrego el componente SafaeAreaView en el componente ListMovies
