+++ src/screens/movie/ScreenMovie.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Text } from "react-native";
-import Movie from "../../components/movie/movie";
+import Movie from "./components/movie/movie";
 import { getDetailsMovie } from "../../services/movies";
 
 export default function ScreenMovie({ route }) {
